refactor spring settings out
